Moved rendering test helpers into their own package
